pyproject.toml
[tool.poetry]
name = "linked_in_posts"
version = "0.0.1"
description = "Code used in my LinkedIn posts"
authors = ["Benjamin Sorenson <benjamin.sorenson81@gmail.com>"]
repository = "https://github.com/jaminsore/linked-in-posts"
documentation = "https://jaminsore.github.io/linked-in-posts/"
readme = "README.md"
packages = [
  {include = "linked_in_posts"}
]

[tool.poetry.dependencies]
python = "^3.11"
fasttext-wheel = "^0.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.12.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
cruft = "^2.15.0"
pytest-jupyter = "^0.7.0"
lorem-text = "^2.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = {extras = ["python"], version = "^0.23.0"}


[tool.poetry.group.repl.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
preview = true

[tool.mypy]
files = ["linked_in_posts", "tests"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[[tool.mypy.overrides]]
module = [
    "fasttext.*",
    "lorem_text.*"
]
ignore_missing_imports = "True"

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["linked_in_posts"]


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
